Removed migrations and converted ints to decimals
Bug: can't turn €10.00 into €10.- while showing €2.50

+++ app/Models/product.php
@@ -12,7 +12,7 @@ class Product extends Model
      * @var array
      */
     protected $casts = [
-        'price' => 'float',
-        'total' => 'float',
+        'price' => 'double',
+        'total' => 'double',
     ];
 }
